Moved additional Linux support from integration test into source
